requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
sentence-transformers
youtube-transcript-api
faiss-cpu